# backend/docs/contracts/loyalty/openapi.yaml
openapi: 3.0.3
info:
  title: CourseFlow Generic Loyalty API
  version: 1.0.0-contract.5
  description: >
    Static OpenAPI contract for the loyalty bounded context. Loyalty owns program configuration,
    account balances, immutable point ledger entries, operational audit, and support workflows.
    Promotion services may emit points intent, but they must not mutate loyalty state directly
    outside loyalty-owned service-to-service commands and consumers.
servers:
  - url: http://localhost:8094
    description: Local loyalty-service
security:
  - internalJwt: []
...
paths:
...
  "/internal/loyalty/dead-letters/{deadLetterId}":
    get:
      tags: [Operations]
      operationId: getLoyaltyInboundDeadLetter
      summary: Get inbound Kafka DLT record detail.
      parameters:
        - name: deadLetterId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Dead-letter detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoyaltyInboundDeadLetterDetail" }
  "/internal/loyalty/dead-letters/{deadLetterId}:replay":
    post:
      tags: [Operations]
      operationId: replayLoyaltyInboundDeadLetter
      summary: Replay an OPEN or FAILED inbound DLT record to its source topic.
      parameters:
        - name: deadLetterId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoyaltyInboundDeadLetterActionRequest" }
      responses:
        "200":
          description: Replay action result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoyaltyInboundDeadLetterActionResponse" }
  "/internal/loyalty/dead-letters/{deadLetterId}:discard":
    post:
      tags: [Operations]
      operationId: discardLoyaltyInboundDeadLetter
      summary: Mark an OPEN or FAILED inbound DLT record as discarded after manual resolution.
      parameters:
        - name: deadLetterId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoyaltyInboundDeadLetterActionRequest" }
      responses:
        "200":
          description: Discard action result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoyaltyInboundDeadLetterActionResponse" }
...
components:
  securitySchemes:
    internalJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by the identity-token-converter for service-to-service calls.
...
  schemas:
...
    LoyaltyInboundDeadLetter:
      type: object
      properties:
        id: { type: string, format: uuid }
        sourceTopic: { type: string }
        dltTopic: { type: string }
        consumerGroup: { type: string, nullable: true }
        kafkaPartition: { type: integer }
        kafkaOffset: { type: integer, format: int64 }
        originalPartition: { type: integer, nullable: true }
        originalOffset: { type: integer, format: int64, nullable: true }
        recordKey: { type: string, nullable: true }
        status: { type: string, enum: [OPEN, FAILED, REPLAYED, DISCARDED] }
        replayAttempts: { type: integer }
        payloadHash: { type: string }
        exceptionClass: { type: string, nullable: true }
        exceptionMessage: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lastReplayAt: { type: string, format: date-time, nullable: true }
        replayedAt: { type: string, format: date-time, nullable: true }
        discardedAt: { type: string, format: date-time, nullable: true }
    LoyaltyInboundDeadLetterDetail:
      allOf:
        - $ref: "#/components/schemas/LoyaltyInboundDeadLetter"
        - type: object
          properties:
            payloadSizeBytes: { type: integer, format: int64 }
            stacktrace: { type: string, nullable: true }
            lastReplayError: { type: string, nullable: true }
            resolvedBy: { type: string, nullable: true }
            resolutionNote: { type: string, nullable: true }
            headers:
              type: object
              additionalProperties: true
...
    LoyaltyInboundDeadLetterActionRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
        dryRun: { type: boolean }
    LoyaltyInboundDeadLetterActionResponse:
      type: object
      properties:
        deadLetterId: { type: string, format: uuid }
        action: { type: string, enum: [REPLAY, DISCARD] }
        status: { type: string, enum: [DRY_RUN, REPLAYED, DISCARDED, FAILED] }
        dryRun: { type: boolean }
        replayed: { type: boolean }
        discarded: { type: boolean }
        reasonCode: { type: string }
        payloadHash: { type: string }
        completedAt: { type: string, format: date-time }
...